chore: Update HelmChart values.yaml to use ClusterIP for internalSmtp type
This code change updates the HelmChart values.yaml file to change the type of the internalSmtp service from LoadBalancer to ClusterIP. The LoadBalancer type was causing issues with receiving emails from the internet, so this change provides a workaround by using the ClusterIP type instead. This update ensures that the internalSmtp service can properly receive emails and improves the functionality of the application.