refactor: Add notIn method to QueryHelper for improved query flexibility
The `notIn` method has been added to the `QueryHelper` class in `Database/QueryHelper.ts`. This method allows for querying records that are not in a specified list of values. It accepts an array of strings or ObjectIDs and generates a `NOT IN` clause for the query. This enhancement provides greater flexibility in constructing complex queries and improves the overall query capabilities of the application.