refactor: Add checkReadBlockPermission method to ReadPermission
This code change adds the checkReadBlockPermission method to the ReadPermission class in ReadPermission.ts. The method is responsible for checking the block permissions for reading records. It allows the query to run if the system is making the query or if the user has root or master admin privileges. This enhancement provides better control over the reading of records and improves the security of the application.