```text refactor: Enable external Postgres support in Helm chart
This code change modifies the Helm chart to enable support for using an external Postgres database. The `postgresql.enabled` flag in the `values.yaml` file is set to `true` to indicate that an external Postgres database will be used. This enhancement allows users to configure the Helm chart to connect to an existing Postgres database instead of deploying a new one.