refactor: Update Postgres configuration in Helm chart
This code change updates the Postgres configuration in the Helm chart to support both internal and external Postgres databases. The `DATABASE_HOST`, `DATABASE_PORT`, `DATABASE_USERNAME`, `DATABASE_PASSWORD`, and `DATABASE_DATABASE` environment variables are now conditionally set based on the `postgresql.enabled` flag in the `values.yaml` file. If `postgresql.enabled` is `true`, the values are set to the internal Postgres service details. Otherwise, they are set to the corresponding values from the `externalPostgres` configuration. This enhancement allows users to easily switch between using an internal or external Postgres database when deploying the Helm chart.