```text refactor: Update FilterViewer component to use initialFilterData prop
This code change updates the FilterViewer component to utilize the initialFilterData prop. If provided, the prop will be used as the initial state for the filterData state variable. This change allows for pre-populating the filter data when rendering the component, providing more flexibility and control over the initial state of the filter.