feat: Add MetricsAggregationType enum and MetricsQuery interface
This commit adds the MetricsAggregationType enum to represent different types of metrics aggregations such as max, min, sum, avg, and count. It also introduces the MetricsQuery interface, which defines the structure of a metrics query including the metric name, attributes, aggregation type, aggregate by, start time, and end time. These changes enable the implementation of metrics queries and aggregations in the application.