refactor: Update email template and resource handling in StatusPageSubscriberReport
The code changes in the `StatusPageSubscriberReport.hbs` template file update the rendering of the resource table based on whether there are any resources to display. If there are resources, the table is rendered with the resource details. If there are no resources, a message is displayed indicating that no resources have been added yet.

In addition, the `SendReportsToSubscribers.ts` file in the `Workers/Jobs/StatusPage` directory has been modified to handle the case where there are no status page resources. If there are no resources, the email report is not sent and an error is logged.

These changes improve the handling of resources in the status page subscriber report and ensure that the email report is only sent when there are resources to include.