refactor: Add doNotFollowRedirects option to WebsiteRequest and API
This commit adds the `doNotFollowRedirects` option to the `WebsiteRequest` and `API` classes. When set to `true`, the classes will not follow redirects when making HTTP requests. This is achieved by setting the `maxRedirects` property of the `axiosOptions` object to 0. This change allows for more control over the behavior of HTTP requests and ensures that redirects are not automatically followed.